requirements.txt
ABBYY==0.3
absl-py==0.9.0
altgraph==0.17
amqp==2.5.2
ansible==2.9.5
anytree==2.8.0
asgiref==3.3.1
astor==0.8.1
astroid==2.0.4
autocommand==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.8.2
billiard==3.6.2.0
blis==0.4.1
boto==2.49.0
boto3==1.12.4
botocore==1.15.4
bottle==0.12.21
cachetools==4.0.0
catalogue==1.0.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
Cython==0.29.15
dataclasses==0.8
datefinder==0.7.1
decorator==4.4.1
deprecation==2.0.7
dj-database-url==0.5.0
Django==3.1.4
django-bootstrap4==2.2.0
django-cors-headers==3.3.0
django-crispy-forms==1.7.2
django-extra-fields==2.0.3
django-extra-fields-lab==0.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-json-field==0.5.7
django-jsonfield==1.4.1
django-mongodb-engine==0.6.0
django-mptt==0.11.0
django-mysql==3.10.0
django-pandas==0.6.1
django-post-request-task==0.4.1
django-restframework==0.0.1
django-sendgrid-v5==1.2.0
django-toolbox==0.2
django-utils==0.0.2
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
djangotoolbox==1.8.0
djongo==1.3.3
docopt==0.6.2
docutils==0.16
en-core-web-sm-abd==0.1
evdev==1.3.0
ffmpy==0.2.2
fiscalyear==0.2.0
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.3.3
gensim==3.8.1
gevent==20.9.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-images-download==2.8.0
Google-Images-Search==1.0.1
google-pasta==0.1.8
greenlet==0.4.17
grpcio==1.27.2
gsmtpd==0.2
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==19.9.0
h5py==2.10.0
httplib2==0.17.0
icrawler==0.6.2
idna==2.9
imageio==2.8.0
importlib-metadata==1.5.0
importlib-resources==3.3.0
imutils==0.5.3
inflection==0.3.1
isort==4.3.4
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
jsonfield==3.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyboard==0.13.5
kiwisolver==1.1.0
kombu==4.6.7
lazy-object-proxy==1.3.1
lxml==4.5.0
mapper==0.1.17
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mock==4.0.1
mongo-connector==3.1.1
mongoengine==0.21.0
more-itertools==8.2.0
MouseInfo==0.1.3
mpyg321==0.0.2
murmurhash==1.0.2
mysql-connector-python==8.0.22
mysqlclient==2.0.3
networkx==2.4
nltk==3.4.5
num2words==0.5.10
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.1.0
packaging==20.1
pandas==1.0.1
paramiko==2.7.2
parsedatetime==2.6
pdf2image==1.12.1
Pillow==6.2.2
pkg-resources==0.0.0
plac==1.1.3
preshed==3.0.2
protobuf==3.11.3
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyaes==1.6.1
pyaml==19.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.19
pyfiglet==0.8.post1
pygame==1.9.6
PyGetWindow==0.0.9
Pygments==2.5.2
PyInstaller==3.6
PyJWT==2.4.0
pymongo==3.11.2
PyMsgBox==1.0.9
PyMySQL==0.9.3
PyNaCl==1.4.0
pynput==1.7.2
pyparsing==2.4.6
PyPDF2==1.26.0
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.2
python-crfsuite==0.9.6
python-dateutil==2.8.1
python-decouple==3.1
python-http-client==3.3.7
python-resize-image==1.1.19
python-util==1.2.1
python-xlib==0.29
python3-xlib==0.15
pytube==9.3.3
PyTweening==1.0.3
pytz==2018.7
PyWavelets==1.1.1
PyYAML==5.3
Quandl==3.5.0
regex==2020.6.8
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
schedule==0.6.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
selenium==3.141.0
sendgrid==6.9.7
six==1.11.0
sklearn==0.0
sklearn-crfsuite==0.3.6
smart-open==1.9.0
smtp==0.1.0
snips-nlu==0.20.2
snips-nlu-parsers==0.4.3
snips-nlu-utils==0.9.1
sounddevice==0.3.14
soupsieve==1.9.5
spacy==2.2.3
SpeechRecognition==3.8.1
SQLAlchemy==1.3.22
sqlparse==0.2.4
srsly==2.0.0
starkbank-ecdsa==2.0.3
stop-words==2018.7.23
tabulate==0.8.6
telebot==0.0.4
Telethon==0.19
tensorboard==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
textblob==0.15.3
tflearn==0.3.2
thinc==7.3.1
tifffile==2020.9.3
tqdm==4.43.0
twilio==7.9.2
typed-ast==1.5.1
uritemplate==3.0.1
urllib3==1.25.8
utils==1.0.1
vine==1.3.0
virtualenv==16.0.0
wasabi==0.6.0
Werkzeug==1.0.0
whitenoise==4.1
wikipedia==1.4.0
wrapt==1.10.11
xlrd==1.2.0
zipp==3.0.0
zope.event==4.5.0
zope.interface==5.2.0
